% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_credit_to_list_of_plot}
\alias{add_credit_to_list_of_plot}
\title{Add credit text to a list of ggplot2 plot object}
\usage{
add_credit_to_list_of_plot(list_of_plot)
}
\description{
Add credit text to a list of ggplot2 plot object
}
