% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_line.R
\name{add_credit_to_plot}
\alias{add_credit_to_plot}
\title{Add credit text to a ggplot2 plot object}
\usage{
add_credit_to_plot(plot_object)
}
\description{
Add credit text to a ggplot2 plot object
}
