% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{change_data_types_of_iamc_dataframe}
\alias{change_data_types_of_iamc_dataframe}
\title{change column data type in dataset to be able to
be treated as an IAMC dataset.}
\usage{
change_data_types_of_iamc_dataframe(iamc_data)
}
\description{
change column data type in dataset to be able to
be treated as an IAMC dataset.
}
