% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_column_availability}
\alias{check_column_availability}
\title{check if the dataset has required fields of IAMC dataset}
\usage{
check_column_availability(iamc_data)
}
\description{
check if the dataset has required fields of IAMC dataset
}
