% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{correct_format_of_iamc_dataframe}
\alias{correct_format_of_iamc_dataframe}
\title{correct data format of given IAMC data table}
\usage{
correct_format_of_iamc_dataframe(iamc_data)
}
\arguments{
\item{iamc_data}{IAMC dataset described above}
}
\value{
modified dataframe
}
\description{
Dataset in IAMC format rule is not rigid.
This function corrects data types of columns in the dataset.
If necessary columns is missing, it throws exception.
Output object of this function is as follows:

type: data.table
columns:
  model:    factor
  scenario: factor
  region:   factor
  variable: factor
  unit:     factor
  period:   double
  value:    double
}
