% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_area.R
\name{generate_code_to_plot_area}
\alias{generate_code_to_plot_area}
\title{generate code to reproduce area plot}
\usage{
generate_code_to_plot_area(input, name_of_input_data_variable,
  name_of_input_rule_table_variable)
}
\arguments{
\item{input}{This is the same as the input argument in the shiny:ui().}

\item{name_of_input_data_variable}{A string such as "ar5_sample_data".}

\item{name_of_input_rule_table_variable}{A string such as "ar5_sample_rule".}
}
\description{
This function is called in the mipplot_interactive_area()
and provides R code to reproduce the currently drawn plot.
This function cannot be used out of reactive expression in Shiny.
}
