% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_line.R
\name{generate_code_to_plot_line}
\alias{generate_code_to_plot_line}
\title{generate code to reproduce line plot}
\usage{
generate_code_to_plot_line(input, name_of_iamc_data_variable = "D")
}
\arguments{
\item{input}{it is same as the argument of shiny::ui()
this function accesses following attributes:
- model
- period
- variable
- scenario
- region}
}
\description{
from `input` argument generally used in
reactive context in Shiny, this function generates
R code to reproduce current plot.
This function could not used out of reactive expression in Shiny.
}
