% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_line.R
\name{get_model_name_list}
\alias{get_model_name_list}
\title{Get name list of models in IAMC formatted data frame}
\usage{
get_model_name_list(D)
}
\arguments{
\item{D}{A quitte format dataframe of IAMC data to produce garph.}
}
\description{
select name of models from the column "model" then make unique it.
output is character vector such as,
c("AIM-Enduse 12.1", "GCAM 3.0", "IMAGE 2.4" )
}
\examples{
\dontrun{
get_model_name_list(ar5_db_sample_data)
}
}
