% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_line.R
\name{get_scenario_name_list}
\alias{get_scenario_name_list}
\title{Get name list of scenarios in IAMC formatted data frame}
\usage{
get_scenario_name_list(D)
}
\arguments{
\item{D}{A quitte format dataframe of IAMC data to produce garph.}
}
\description{
select name of scenarios from the column "scenario" then make unique it.
output is character vector such as,
c("EMF27-450-Conv", "EMF27-450-FullTech", "EMF27-450-NoCCS", "EMF27-450-NucOff")
}
\examples{
\donttest{
get_scenario_name_list(ar5_db_sample_data)
}
}
