% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_string_expression_of_vector_of_strings}
\alias{get_string_expression_of_vector_of_strings}
\title{Get expression of vector of string in string format}
\usage{
get_string_expression_of_vector_of_strings(vector_of_strings)
}
\arguments{
\item{vector_of_strings}{vector of strings, such as c("A", "B")}
}
\value{
An R code representing character vector
}
\description{
To evaluate expression, get string of expression
}
