% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_variable_group_name_list}
\alias{get_variable_group_name_list}
\title{Get variable-group-name list}
\usage{
get_variable_group_name_list(rule_table)
}
\arguments{
\item{rule_table}{A rule table}
}
\value{
variable group name
}
\description{
variable-group is a combination of one LHS and one or more RHS.
this function outputs the list of names of variable-group in given rule-table.
the format of return value is "LHS|RHS1,RHS2,RHS3,...".
}
\examples{
get_variable_group_name_list(ar5_db_sample_rule_table)
}
