% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_variable_name_list_in_variable_group}
\alias{get_variable_name_list_in_variable_group}
\title{Get variable name list in given variable-group}
\usage{
get_variable_name_list_in_variable_group(group_name)
}
\arguments{
\item{group_name}{variable-group-name}
}
\value{
A list of strings representing variable names
}
\description{
Scan rule-table and extract variable names in given variable-group.
}
\examples{
get_variable_name_list_in_variable_group(
  "Final Energy|Industry,Residential and Commercial,Transportation")
}
