% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot.R
\docType{package}
\name{mipplot}
\alias{mipplot}
\alias{mipplot-package}
\title{mipplot}
\description{
Package contains generic functions to produce area/bar/box/line plots of data following IAMC submission format.
}
