% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_bar_v02.R
\name{mipplot_bar}
\alias{mipplot_bar}
\title{A function to plot bar graph}
\usage{
mipplot_bar(D, R, region = levels(D$region),
  target_year = levels(as.factor(D$period)), facet_x = NULL,
  facet_y = NULL, PRINT_OUT = F, DEBUG = T, fontsize = 20)
}
\arguments{
\item{D}{A dataframe of IAMC data to produce garph.}

\item{R}{A dataframe of data aggregation rules (meta data).}
}
\value{
A list of graph
}
\description{
Bar plots using right-hand-side values of target additivity rule.
}
