% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_generate_color_mapper.R
\name{mipplot_generate_color_mapper}
\alias{mipplot_generate_color_mapper}
\title{Manual coloring}
\usage{
mipplot_generate_color_mapper(raw_table, category_separator = "\\\\|")
}
\arguments{
\item{raw_table}{rule table which includes "Color_code" column.}

\item{category_separator}{regular expression for separating
right-hand-side variable name into categories.
For example: separator should be "\\|" for "Secondary Energy|Electricity|Coal"}
}
\value{
named list of named string vectors.
for example,

  result = list(
    "Emissions|CO2" = c(
      "Fossi Fuels and Industry" = "#17202a",
      "Land Use" = "#008000", ...),
    "Emissions|CO2|Fossil Fuels and Industry" = c(
      "Energy Demand" = "#444444", ...
    ),...
}
\description{
Generate mapper from name of variable to name of color
}
