% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_bar.R
\name{mipplot_interactive_bar}
\alias{mipplot_interactive_bar}
\title{A function to launch interactive plot using Shiny}
\usage{
mipplot_interactive_bar(D, R)
}
\arguments{
\item{D}{A quitte format dataframe of IAMC data to produce garph.}

\item{R}{A table with additivity rules.}
}
\description{
A function to launch interactive bar plot using
             right-hand-side values of  target additivity rule.
             The function arguments include the input dataframe,
             labels for the plot/axes/legend, and faceting dimensions
}
\examples{
\donttest{
mipplot_interactive_bar(ar5_db_sample_data, ar5_db_sample_rule_table)
}
}
