% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_line.R
\name{mipplot_interactive_line}
\alias{mipplot_interactive_line}
\title{A function to launch interactive plot using Shiny}
\usage{
mipplot_interactive_line(D, language = "en")
}
\arguments{
\item{D}{A quitte format dataframe of IAMC data to produce garph.}

\item{language}{A string of language for initial plot.
Possible values are "en", "jp",
"es", "zh-cn", "zh-tw". The default value is "en".}
}
\description{
A function to launch interactive line plot.
             The function arguments include the input dataframe,
             labels for the plot/axes/legend, and faceting dimensions
}
\examples{
\donttest{
mipplot_interactive_line(ar5_db_sample_data)
}
}
