% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_point_v01.R
\name{mipplot_point}
\alias{mipplot_point}
\title{A function to plot point graph}
\usage{
mipplot_point(D, region = levels(D$region), variable = levels(D$variable),
  target_year = levels(D$period), PRINT_OUT = F, DEBUG = T)
}
\arguments{
\item{D}{A dataframe of IAMC data to produce garph.}
}
\value{
A list of graph
}
\description{
point plots
}
