% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_print_pdf.R
\name{mipplot_print_pdf}
\alias{mipplot_print_pdf}
\title{Print list of plots to pdf file}
\usage{
mipplot_print_pdf(
  p_list1,
  filelabel = "",
  filename = tryCatch(file.choose(new = TRUE), error = function(e) {     NA })
)
}
\arguments{
\item{p_list1}{A list of ggplot plot.}

\item{filelabel}{A string of prefix of output filename.}

\item{filename}{A string of filename. If it is given,
filelabel is ignored.}
}
\value{
No return value, called for side effects
}
\description{
This function plots a ggplot plots to PDF file.
}
\examples{
\donttest{
if (interactive()) {
p <- mipplot_area(ar5_db_sample_data, ar5_db_sample_rule_table,
             region = "World", scenario = "EMF27-450-FullTech")
mipplot_print_pdf(p)
}
}
}
