% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_read_iamc.R
\name{mipplot_read_iamc}
\alias{mipplot_read_iamc}
\title{Read IAMC scenario input data.}
\usage{
mipplot_read_iamc(
  filename = NULL,
  sep = ",",
  interactive = FALSE,
  DEBUG = TRUE
)
}
\arguments{
\item{filename}{Path to a file containing scenario data in IAMC format.}

\item{sep}{A character indicating the separator used in the input file.}

\item{interactive}{open a dialog for selecting file if interactive=TRUE.}

\item{DEBUG}{experimental.}
}
\value{
A dataframe in tibble format ("model, scenario, variable, unit, period, value")
}
\description{
Read scenario input data (in IAMC format) as tibble format dataframe.
}
\examples{
\dontrun{
mipplot_read_iamc("filename")
}
}
