% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_readquitte.R
\name{mipplot_readquitte}
\alias{mipplot_readquitte}
\title{Read IAMC scenario input data.}
\usage{
mipplot_readquitte(filename = NULL, sep = ",", interactive = FALSE,
  DEBUG = T)
}
\arguments{
\item{filename}{Path to a file containing scenario data in IAMC format.}

\item{sep}{A character indicating the separator used in the input file.}

\item{interactive}{open a dialog for selecting file if interactive=TRUE.}

\item{DEBUG}{experimental.}
}
\value{
A dataframe in quitte format ("model, scenario, variable, unit, period, value")
}
\description{
Read scenario input data (in IAMC format) as quitte format dataframe.
}
\examples{
\donttest{
mipplot_readquitte("filename")
}
}
