% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_return_table_github.R
\name{mipplot_return_table}
\alias{mipplot_return_table}
\title{Mutated table of SR15 Data}
\usage{
mipplot_return_table(D, R)
}
\arguments{
\item{D}{A dataframe of IAMC data in tibble format to produce mutated table}

\item{R}{A dataframe of data aggregation rules}
}
\value{
Mutated Table of model,scenario,region,variable,unit,period,value
}
\description{
Mutated Table using filtered variable from the rule table
         The function arguments include the input dataframes:
         The SR15 dataset and the Rule Table and returns a mutated
         table with variable, value, model, scenario, region, period
}
\examples{
mipplot_return_table(sr15_sample_data, sr15_sample_conversion_rule_table)
}
