% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_var_submission.R
\name{mipplot_var_submission}
\alias{mipplot_var_submission}
\title{variable SUBMISSION CHECK}
\usage{
mipplot_var_submission(D, V, na_name = "N/A")
}
\arguments{
\item{D}{input data table}

\item{V}{list of variables}

\item{na_name}{string for N/A}
}
\description{
Verify whether data of variables included
 in list template have been submitted.
}
