% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_read_iamc.R
\name{read_iamc_xlsx}
\alias{read_iamc_xlsx}
\title{Read IAMC scenario input data in Excel format}
\usage{
read_iamc_xlsx(file_path, sheet = 2)
}
\arguments{
\item{file_path}{Path to a file containing scenario data in IAMC format.}

\item{sheet}{the index of sheet which contains records.}
}
\value{
A dataframe in tibble format ("model, scenario, variable, unit, period, value")
}
\description{
Read scenario input data (in IAMC format) as tibble format dataframe from Excel
}
\examples{
\dontrun{
read_iamc_xlsx("c:\\\\...\\\\...", sheet = 2)
}
}
