% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{split_variable_into_positive_and_negative_parts}
\alias{split_variable_into_positive_and_negative_parts}
\title{Split variable into positive and negative parts}
\usage{
split_variable_into_positive_and_negative_parts(
  df_all,
  domain_column_name,
  variable_column_name,
  value_column_name,
  variable_name_converter = function(x) {     paste(x, "_negative", sep = "") },
  increment_of_domain_in_interpolation = 0.1
)
}
\value{
modified data frame
}
\description{
Split variable into positive and negative parts
}
