% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sr15_sample_conversion_rule_table.R
\docType{data}
\name{sr15_sample_conversion_rule_table}
\alias{sr15_sample_conversion_rule_table}
\title{Sample Conversion Rule Table}
\format{
An object of class \code{data.frame} with 37 rows and 6 columns.
}
\usage{
sr15_sample_conversion_rule_table
}
\description{
A sample conversion rule table for mipplot_return_table.
}
\keyword{datasets}
