% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sr15_sample_data.R
\docType{data}
\name{sr15_sample_data}
\alias{sr15_sample_data}
\title{Sample Dataset}
\format{A tibble data.table with 396425 rows and 7 variables:
\describe{
  \item{model}{model, categorical}
  \item{scenario}{scenario, categorical}
  \item{region}{region, ASIA, OECD90 or World}
  \item{variable}{the name of simulated variable that changes over time}
  \item{unit}{unit of a variable}
  \item{period}{year}
  \item{value}{the value of a variable}
}}
\usage{
sr15_sample_data
}
\description{
A sample dataset of IAMC format consist of a subset of IPCC special report (Global Warming of 1.5°C, 2018).
}
\keyword{datasets}
