% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{add_credit_to_plot}
\alias{add_credit_to_plot}
\title{Add credit text to a plot}
\usage{
add_credit_to_plot(plot_object)
}
\arguments{
\item{plot_object}{ggplot2 plot object}
}
\value{
modified ggplot2 plot object
}
\description{
Add credit text and project URL to a ggplot2 plot object
}
