% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_column_availability}
\alias{check_column_availability}
\title{check if the dataset has required fields of IAMC dataset}
\usage{
check_column_availability(iamc_data)
}
\arguments{
\item{iamc_data}{IAMC data frame}
}
\value{
boolean flag
}
\description{
if dataset has all required fields, then returns TRUE
}
