% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_format_of_iamc_dataframe}
\alias{check_format_of_iamc_dataframe}
\title{check if the format of given data is valid
as an IAMC dataset.}
\usage{
check_format_of_iamc_dataframe(iamc_data)
}
\arguments{
\item{iamc_data}{IAMC dataset in dataframe format}
}
\value{
TRUE if it is valid
}
\description{
check if the format of given data is valid
as an IAMC dataset.
}
