% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_variable_name_list_in_variable_group}
\alias{get_variable_name_list_in_variable_group}
\title{Get variable name list in given variable-group}
\usage{
get_variable_name_list_in_variable_group(group_name)
}
\arguments{
\item{group_name}{variable-group-name}
}
\description{
Scan rule-table and extract variable names in given variable-group.
}
\examples{
\dontrun{
noquote(
  get_variable_name_list_in_variable_group(
    ar5_db_sample_rule_table,
    "Final Energy|Industry,Residential and Commercial,Transportation")
)
}
}
