% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_area_v03.R
\name{mipplot_area}
\alias{mipplot_area}
\title{A function to plot area graph}
\usage{
mipplot_area(D, R, region = levels(D$region), scenario = levels(D$scenario),
  facet_x = NULL, facet_y = NULL, PRINT_OUT = F, DEBUG = T,
  fontsize = 20)
}
\arguments{
\item{D}{A dataframe of IAMC data to produce area plot.}

\item{R}{A dataframe of data aggregation rules (meta data).}
}
\value{
A list of area plots
}
\description{
Area plots using right-hand-side values of target
  additivity rule. The function arguments inlcude the input dataframe,
  labels for the plot/axes/legend, and faceting dimensions (two in this version).
}
