% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_area.R
\name{mipplot_area}
\alias{mipplot_area}
\title{Area plot from IAMC data}
\usage{
mipplot_area(
  D,
  R,
  region = levels(D$region),
  scenario = levels(D$scenario),
  facet_x = NULL,
  facet_y = NULL,
  PRINT_OUT = F,
  DEBUG = T,
  fontsize = 20,
  color_code_specify = T,
  one_hundred_percent_stacked = F,
  language = "en"
)
}
\arguments{
\item{D}{A dataframe of IAMC data in tibble format to produce area plots.}

\item{R}{A dataframe of data aggregation rules (meta data).}

\item{region}{A list of regions.}

\item{scenario}{A list of scenario.}

\item{facet_x}{facet_x}

\item{facet_y}{facet_y}

\item{PRINT_OUT}{set TRUE to generate PDF file.}

\item{DEBUG}{set TRUE to show debug messages.}

\item{fontsize}{font size of text.}

\item{color_code_specify}{set FALSE if you apply default color palette.}

\item{one_hundred_percent_stacked}{set TRUE if you want a graph of 100% stacked, set this to TRUE.}

\item{language}{A string of language. Possible values are "en", "jp",
"es", "zh-cn", "zh-tw". The default value is "en".}
}
\value{
A list of area plots.
}
\description{
Area plots using right-hand-side values of target
             additivity rule. The function arguments include the input dataframe,
             labels for the plot/axes/legend, and faceting dimensions
             (two in this version).
}
\examples{
\dontrun{
mipplot_area(ar5_db_sample_data, ar5_db_sample_rule_table)
}
}
