% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_box_v02.R
\name{mipplot_box}
\alias{mipplot_box}
\title{A function to plot box graph}
\usage{
mipplot_box(D, region = levels(D$region), variable = levels(D$variable),
  target_year = levels(D$period), PRINT_OUT = F, DEBUG = T)
}
\arguments{
\item{D}{A dataframe of IAMC data to produce graph.}
}
\value{
A list of graph
}
\description{
box plots
}
