% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_mipplot_default_color_palette.R
\docType{data}
\name{mipplot_default_color_palette}
\alias{mipplot_default_color_palette}
\title{Default color palette.}
\format{An object of class \code{character} of length 23.}
\usage{
mipplot_default_color_palette
}
\description{
A default color palette object,
which maps variable name (such as "Land Use")
to hex color code.
}
\keyword{datasets}
