% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_generate_color_mapper.R
\name{mipplot_generate_color_mapper}
\alias{mipplot_generate_color_mapper}
\title{Manual coloring}
\usage{
mipplot_generate_color_mapper(raw_table, category_separator = "\\\\|")
}
\arguments{
\item{raw_table}{rule table which includes "Color_code" column.}

\item{category_separator}{regular expression for separating
right-hand-side variable name into categories.
For example: separator should be "\\|" for "Secondary Energy|Electricity|Coal"}
}
\value{
list. which elements are mapping from variable name to color
}
\description{
Generate mapper from name of variable to name of color
}
