% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_additivity_check_bar.R
\name{mipplot_interactive_additivity_check_bar}
\alias{mipplot_interactive_additivity_check_bar}
\title{A function to launch interactive plot using Shiny}
\usage{
mipplot_interactive_additivity_check_bar(D, R, debug = FALSE)
}
\arguments{
\item{D}{A quitte format dataframe of IAMC data to produce graph.}

\item{R}{A table with additivity rules.}

\item{debug}{Set TRUE if table view is required.}
}
\description{
A function to launch interactive plot for additivity check.
}
\examples{
\dontrun{
mipplot_interactive_additivity_check_bar(ar5_db_sample_data, ar5_db_sample_rule_table)
}
}
