% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_area.R
\name{mipplot_interactive_area}
\alias{mipplot_interactive_area}
\title{A function to launch interactive plotting session on Shiny}
\usage{
mipplot_interactive_area(D, R, language = "en")
}
\arguments{
\item{D}{A dataframe of IAMC data in tibble format to produce area plots.}

\item{R}{A dataframe of data aggregation rules (meta data).}

\item{language}{A string of language for initial plot.
Possible values are "en", "jp",
"es", "zh-cn", "zh-tw". The default value is "en".}
}
\description{
Provides gui to set plotting parameter for area plot.
}
\examples{
\donttest{
if (interactive()) {
mipplot_interactive_area(ar5_db_sample_data, ar5_db_sample_rule_table)
}
}
}
