% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_interactive_plot_line.R
\name{mipplot_interactive_plot_line}
\alias{mipplot_interactive_plot_line}
\title{A function to launch interactive plot using Shiny}
\usage{
mipplot_interactive_plot_line(D, R)
}
\arguments{
\item{D}{A quitte format dataframe of IAMC data to produce garph.}

\item{R}{A table with additivity rules.}
}
\description{
A function to launch interactive plot using Shiny
}
