% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_line_v02.R
\name{mipplot_line}
\alias{mipplot_line}
\title{A function to plot line graph}
\usage{
mipplot_line(D, region = levels(D$region), variable = levels(D$variable),
  colorby = "scenario", linetypeby = "model", shapeby = "model",
  scenario = levels(D$scenario), facet_x = NULL, facet_y = NULL,
  PRINT_OUT = F, DEBUG = T)
}
\arguments{
\item{D}{A dataframe of IAMC data to produce garph.}
}
\value{
A list of plots.
}
\description{
Line plots
}
