% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_line.R
\name{mipplot_line}
\alias{mipplot_line}
\title{Line plot from IAMC data}
\usage{
mipplot_line(D, region = levels(D$region),
  variable = levels(D$variable), colorby = "scenario",
  linetypeby = "model", shapeby = "model",
  scenario = levels(D$scenario), facet_x = NULL, facet_y = NULL,
  PRINT_OUT = F, DEBUG = T, language = "en")
}
\arguments{
\item{D}{A dataframe of IAMC data in tibble format to produce plots.}

\item{region}{A list of regions.}

\item{variable}{A list of variables.}

\item{colorby}{an axis for color setting.}

\item{linetypeby}{an axis for line type setting.}

\item{shapeby}{an axis for shape setting.}

\item{scenario}{A list of cenarios.}

\item{facet_x}{facet_x}

\item{facet_y}{facet_y}

\item{PRINT_OUT}{set TRUE to generate PDF files.}

\item{DEBUG}{set TRUE to show debug messages.}

\item{language}{A string of language. Possible values are "en", "jp",
"es", "zh-cn", "zh-tw". The default value is "en".}
}
\value{
A list of line plots.
}
\description{
The function arguments include the input dataframe,
             labels for the plot/axes/legend, and faceting dimensions
}
\examples{
\donttest{
mipplot_line(ar5_db_sample_data)
}
}
