% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_read_ruletab.R
\name{mipplot_read_ruletab}
\alias{mipplot_read_ruletab}
\title{A function to read rule table without ID number}
\usage{
mipplot_read_ruletab(R_without_id)
}
\description{
Read table of additivity rule and adds column with id number.
}
