% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipplot_var_submission_v01.R
\name{mipplot_var_submission}
\alias{mipplot_var_submission}
\title{variable SUBMISSION CHECK}
\usage{
mipplot_var_submission(D, V, na_name = "N/A")
}
\description{
Verify whether data of variables included
 in list template have been submitted.
}
